requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.10
Django==5.1.6
django-allauth==65.4.1
django-cors-headers==4.7.0
django-crispy-forms==2.3
djangorestframework==3.15.2
sqlparse==0.5.3
tzdata==2025.1
